Fix -Wsign-compare when -DRUBY_DEBUG=1
Sizes for darray are size_t now. CC @peterzhu2118

+++ yjit_core.c
@@ -567,7 +567,7 @@ add_block_version(block_t *block)
 #endif
     }
 
-    RUBY_ASSERT((int32_t)blockid.idx < rb_darray_size(body->yjit_blocks));
+    RUBY_ASSERT(blockid.idx < rb_darray_size(body->yjit_blocks));
     rb_yjit_block_array_t *block_array_ref = rb_darray_ref(body->yjit_blocks, blockid.idx);
 
     // Add the new block
